refactor: fix placements
