+++ src/ts/status/signin/actions/use-signin.test.tsx
@@ -6,7 +6,7 @@ import createUseSignIn from "./use-signin";
 import React, { useEffect } from "react";
 import { render } from "@testing-library/react";
 import { flushPromisesAndTimers, RecoilObserver } from "@/lib.test";
-import currentUserState from "../atoms/current-user";
+import currentUserState from "@/status/user/atoms/current-user-state";
 
 test("do not update state if user is not found", async () => {
   const registrar = createDependencyRegistrar<Dependencies>();
