+++ src/ts/status/signin/atoms/signin-state.ts
@@ -1,12 +1,13 @@
 import { atom } from "recoil";
 import AtomKeys from "./key";
 
-const signInState = atom<{ password: string; email: string; authenticating: boolean }>({
+const signInState = atom<{ password: string; email: string; authenticating: boolean; authenticated: boolean }>({
   key: AtomKeys.signIn,
   default: {
     email: "",
     password: "",
     authenticating: false,
+    authenticated: false,
   },
 });
