+++ src/ts/status/signin/selectors/authenticated-state.ts
@@ -1,11 +1,11 @@
 import { selector } from "recoil";
-import currentUserState from "../atoms/current-user";
+import signInState from "../atoms/signin-state";
 import SelectorKeys from "./key";
 
 const authenticatedState = selector<boolean>({
   key: SelectorKeys.authenticated,
   get: ({ get }) => {
-    return get(currentUserState).id !== null;
+    return get(signInState).authenticated;
   },
 });
