+++ src/ts/status/signin/selectors/authenticating-state.ts
@@ -1,5 +1,5 @@
 import { selector } from "recoil";
-import signInState from "../atoms/signing";
+import signInState from "../atoms/signin-state";
 import SelectorKeys from "./key";
 
 const authenticatingState = selector<boolean>({
